# dist/api.yml
openapi: 3.0.0
paths: 
...
  /users/favoriteGames:
     get:
      tags:
        - "User"
      summary: "Retrieve the favorite games for the logged in user"
      description: "The favorite games are upcoming games and so they won't have an evenlog and a result (both will be null).
      If the user does not have favorite games the array of games will be empty"
      operationId: "getFavGamesForUser" 
      responses:
        "404":
          description: "URL not found"
        "401":
          description: "Unauthorized, please login first"
        "200":
          description: "Game was added to favorite"
          content:  
            application/json:    
              schema:            
                $ref: "#/components/schemas/ArrayOfFutureGames"
  /users/addFavoriteGames:
      post:
        tags:
          - "User"
        description: "add the favorite game to the list of the user's favorite games"
        summary: "used in the user's personal page"
        requestBody:
          content:
            application/json:  
              schema:
                type: object
                properties:
                  game_id:          
                    type: "integer"
                    example: 1   

        responses:
          "404":
            description: "URL not found"
          "403":
            description: "Can only add future games to favorite games list"
          "401":
            description: "Unauthorized, please login first" 
          "400":
            description: "Invalid Date or Teams supplied"
          "200":
            description: "Game was added succesfully"
  /users/favoriteTeams:
    get:
      tags:
        - "User"
      summary: "Retrieve the favorite teams for the logged in user"
      description: "Retrieve the favorite teams the user marked. If there are none, an empty array will return."
      operationId: "getFavTeamsForUser" 
          
      responses:
        "404":
          description: "URL not found"
        "401":
          description: "Unauthorized, please login first"
        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:            
                $ref: "#/components/schemas/ArrayOfFavoriteTeams"
  /users/addFavoriteTeams:
    post:
      tags:
        - "User"
      description: "add the favorite team to the list of the user's favorite teams."
      summary: "used in the user's personal page"
      requestBody:
        content:
          application/json:  
            schema:
              type: object
              properties:
                team_id:
                  type: "integer"
                  example: 939
          
      responses:
        "404":
          description: "URL not found"
        "401":
          description: "Unauthorized, please login first"
        "400":
          description: "Invalid team id"
        "200":
          description: "Team was added succesfully"
  /users/favoritePlayers:
    get:
      tags:
        - "User"
      summary: "Retrieve the favorite players for the logged in user"
      description: "Retrive all the favorite players that the user marked. If there are none, an empty array will return."
      operationId: "getFavPlayersForUser" 
          
      responses:
        "404":
          description: "URL not found"
        "401":
          description: "Unauthorized, please login first"
        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:            
                $ref: "#/components/schemas/ArrayOfPlayers"
  /users/addFavoritePlayers:
    post:
      tags:
        - "User"
      description: "add the favorite player to the list of the user's favorite players"
      summary: "used in the user's personal page"
      requestBody:
        content:
          application/json:  
            schema:
              type: object
              properties:
                player_id:
                  type: "integer"
                  example: 172104
          
      responses:
        "404":
          description: "URL not found"
        "401":
          description: "Unauthorized, please login first"
        "400":
          description: "Invalid player id"
        "200":
          description: "player was added succesfully"
...
  /players/search/{player_name}/team/{team_name}:
    get:
      tags:
        - "Search"
      summary: "Used for the search page."
      description: "Get part of the player's information by using the player's fullname and team as filter.
      All the players with that name and team name will return. If there are none, an empty array will return."
      parameters:
      - name: "player_name"
        in: "path"
        description: "Full name of player to return"
        required: true
        schema:
          type: "string"
          example: "Esben Lange"
      - name: "team_name"   
        in: "path"
        description: "Player's team name"
        required: true
        schema:
          type: "string"
          example: "Silkeborg"

      responses:
        "404":
          description: "URL not found"
        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:
                type: array
                items:
                  type: object
                  allOf:
                     - $ref: "#/components/schemas/PlayerPreview"        
  /teams/:
    get:
      tags:
        - "Team"
      description: "Get an array of all the teams in the league"
      summary: "Used to find teams to add to favorite"

            
      responses:
        "404":
          description: "URL not found - could be because of bad id"
        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:            
                type: array
                items:
                  type: object
                  allOf:
                    -  $ref: "#/components/schemas/TeamSummary"
  /teams/id/{teamId}:
    get:
      tags:
        - "Team"
      description: "Get information about the team by using the team id"
      summary: "used in the team's personal page"
      operationId: "getTeamById" 
      parameters:
      - name: "teamId"
        in: "path"
        description: "ID of team to return"
        required: true
        schema:
          type: "integer"
          minimum: 1
          example: 211
            
      responses:
        "404":
          description: "URL not found - could be because of bad id"
        "400":
          description: "Team id not found"
        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:            
                $ref: "#/components/schemas/Team"
  /teams/name/{team_name}:
    get:
      tags:
        - "Team"
      description: "Get information about the team by using the team name"
      summary: "used in the team's personal page"
      operationId: "getTeamByName" 
      parameters:
      - name: "team_name"
        in: "path"
        description: "Name of team to return"
        required: true
        schema:
          type: "string"
          example: "Silkeborg"
            
      responses:
        "404":
          description: "URL not found - could be because of bad name"
        "400":
          description: "Team name not found"
        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:            
                $ref: "#/components/schemas/Team"
  /teams/teamID/{team_name}:
    get:
      tags:
        - "Team"
      description: "Get ID of the team by using the team name"
      summary: "used in the add games page"
      operationId: "getTeamIDByName" 
      parameters:
      - name: "team_name"
        in: "path"
        description: "Name of team id to return"
        required: true
        schema:
          type: "string"
          example: "Silkeborg"
            
      responses:
        "404":
          description: "URL not found - could be because of bad name"
        "400":
          description: "Team name not found"
        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:
                type: object
                properties:
                  team_id:
                    type: "integer"
                    example: 939
  /teams/search/{team_name}:
    get:
      tags:
        - "Search"
      summary: "Used for the search page."
      description: "Get part of the team's information by using the team's name. If there are none, an empty array will return."
      parameters:
      - name: "team_name"
        in: "path"
        description: "Team name of the team to return"
        required: true
        schema:
          type: "string"
          example: "Silkeborg"
            
      responses:
        "404":
          description: "URL not found"
        "400":
          description: "Invalid team name"
        "200":
          description: "successful operation"
          content:  
            application/json:    
              schema:
                type: object
                properties:
                  id:
                    type: "integer"
                    example: 86
                  legacy_id:
                    type: "integer"
                    example: 522
                  name: 
                    type: "string"
                    example: "Silkeborg"
                  short_code:
                    type: "string"
                    example: "SIL"
                  twitter: 
                    type: "string"
                    example: "https://twitter.com/SilkeborgTeam.com"
                  country_id: 
                    type: "integer"
                    example: 320
                  national_team:
                    type: "boolean"
                    example: false                    
                  founded: 
                    type: "integer"
                    example: 1917
                  logo_path:
                    type: "string"
                    example: "https://cdn.sportmonks.com/images//soccer/teams/22/86.png"
                  venue_id:
                    type: "integer"
                    example: 85371
                  current_season_id:
                    type: "integer"                    
                    example: 18334
                  is_placeholder:
                    type: "boolean"
                    example: false
...
components:
  schemas:
...
    TeamSummary:
      type: object
      properties:
        team_id:
          type: integer
          example: 211
        team_name:
          type: "string"
          example: "Renders"     
    Team:
      type: object
      properties:
        id:
            type: "integer"
            format: "int64"
            example: 211
        name:
            type: "string"
            example: "Silkeborg"
        logo:
            type: "string"
            example: "https://cdn.sportmonks.com/images//soccer/teams/11/211.png"
        coach:
            $ref: "#/components/schemas/Coach"
        stadium:
            $ref: "#/components/schemas/Stadium"
        squad:
          type: array
          items:
            type: object
            allOf:
              - $ref: "#/components/schemas/Player"
        latest:
          type: array
          items:
            type: object
            allOf:
              - $ref: "#/components/schemas/Game"
        upcoming:
          type: array
          items:
            type: object
            allOf:
              - $ref: "#/components/schemas/Game"
    FavoriteTeam:
      type: object
      properties:
        id:
            type: "integer"
            format: "int64"
            example: 211
        name:
            type: "string"
            example: "Silkeborg"
        logo:
            type: "string"
            example: "https://cdn.sportmonks.com/images//soccer/teams/11/211.png"
        coach:
            $ref: "#/components/schemas/Coach"
        stadium:
            $ref: "#/components/schemas/Stadium"
        latest:
          type: array
          items:
            type: object
            allOf:
              - $ref: "#/components/schemas/Game"
        upcoming:
          type: array
          items:
            type: object
            allOf:
              - $ref: "#/components/schemas/Game"
    Stadium:
      type: object
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 5653  
        name:
          type: "string"
          example: "Sydbank Park"
        surface:
          type: "string"
          example: "grass"
        address:
          type: "string"
          example: "Stadionvej 7"
        city: 
          type: "string"
          example: "Herning"
        capacity:
          type: "integer"
          format: "int64"
          example: "Haderslev"
        image_path:
          type: "string"
          example: "https://cdn.sportmonks.com/images/soccer/venues/21/5653.png"
        coordinates:
          type: "string"
          example: "55.261402,9.487573"
...
    FutureGame:
      required:
        - game_date
        - game_time
        - home_team
        - away_team
        - stadium
      type: object
      properties:
        gameid:
          type: "integer"
          example: 1
        GameDateTime:
          type: "string"
          example: "2021-10-28 17:00:00"
        HomeTeam:
          type: "string"
          example: "Midtjylland"
        HomeTeamID:
          type: "integer"
          example: 939
        AwayTeam:
          type: "string"
          example: "Silkeborg"
        AwayTeamID:
          type: "integer"
          example: 86
        Stadium:
          type: "string"
          example: "MCH Arena"
        Referee:
          type: "string"
          example: "Denis Shalayev"
    Game:
      required:
        - game_date
        - game_time
        - home_team
        - away_team
        - stadium

      type: object
      properties:
        gameid:
          type: "integer"
          example: 1
        GameDateTime:
          type: "string"
          example: "2021-01-13 19:00:00"
        HomeTeam:
          type: "string"
          example: "Midtjylland"
        HomeTeamID:
          type: "integer"
          example: 939
        AwayTeam:
          type: "string"
          example: "Silkeborg"
        AwayTeamID:
          type: "integer"
          example: 211
        Stadium:
          type: "string"
          example: "MCH Arena"
        Referee:
          type: "string"
          example: "Denis Shalayev"
        Result:
          type: "string"
          example: "3-1"
        event_log:
          $ref: "#/components/schemas/EventLog"
...
    GameEvent:
      description: "An Event in a game. Options are: Red Card, Yellow Card, Foul, Offside, Goal, Injury, Subsitution"
      type: object
      properties:
        EventData:
          type: "string"
          example: "2021-01-13"
        EventTime:
          type: "string"
          example: "20:05:00"
        EventGameTime:
          type: "integer"
          format: "int64"
          example: 65  
        EventDesc:
          type: "string"
          example: "Foul Anders Hoff"
    Player:
      type: object
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 37522564  
        team_id:
          type: "integer"
          format: "int64"
          example: 211  
        position:
          type: "string"
          example: "Midfielder" 
        position_id:
          type: "integer"
          example: 2 
        fullname:
          type: "string"
          example: "James Forest"
        common_name:
          type: "string"
          example: "James Forest"
        nationality:
          type: "string"
          example: "Scotland"  
        birthplace:
          type: "string"
          example: "Prestwick" 
        birthdate:
          type: "string"
          example: "07/07/1991"  
        height:
          type: "string"
          example: "175 cm"  
        weight:
          type: "string"
          example: "90 kg"
        team:
          type: "string"
          example: "Celtic"
        image_path:
          type: "string"
          example: "https://cdn.sportmonks.com/images/soccer/players/8/172104.png"
    PlayerPreview:
      type: object
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 172104  
        team_id:
          type: "integer"
          format: "int64"
          example: 53  
        position:
          type: "string"
          example: "Midfielder" 
        position_id:
          type: "integer"
          example: 2 
        fullname:
          type: "string"
          example: "James Forest"
        common_name: 
          type: "string"
          example: "R. Almenning Jarstein"
        nationality: 
          type: "string"
          example: "Norway"
        birthplace:
          type: "string"
          example: "Norway"
        birthdate: 
          type: "string"
          example: "29/09/1984"
        height:
          type: "string"
          example: "192 cm"
        weight:
          type: "string"
          example: "88 kg"
        team:
          type: "string"
          example: "Celtic"
        image_path:
          type: "string"
          example: "https://cdn.sportmonks.com/images/soccer/players/8/172104.png"
    Coach:
      type: object
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1467946  
        team_id:
          type: "integer"
          format: "int64"
          example: 53  
        country_id:
          type: "integer"
          format: "int64"
          example: 491  
        fullname:
          type: "string"
          example: "Neil Lennon"
        firstname:
          type: "string"
          example: "Neil"
        lastname:
          type: "string"
          example: "Lennon"
        nationality:
          type: "string"
          example: "Northern Ireland"  
        birthplace:
          type: "string"
          example: "Lurgan" 
        birthdate:
          type: "string"
          example: "25/06/1971"  
        image_path:
          type: "string"
          example: "https://cdn.sportmonks.com/images/soccer/players/10/1467946.png"
    EventLog:
      description: "The Eventlog - contains array of all the events in the game. Used only in past games"
      type: array
      items:
        type: object
        allOf:
          - $ref: "#/components/schemas/GameEvent"
...
    ArrayOfFavoriteTeams:
      type: array
      items:
        type: object
        allOf:
          - $ref: "#/components/schemas/FavoriteTeam"
...
    ArrayOfPlayers:
      type: array
      items:
        type: object
        allOf:
          - $ref: "#/components/schemas/Player"
...
    ArrayOfFutureGames:
      type: array
      items:
        type: object
        allOf:
          - $ref: "#/components/schemas/FutureGame"